added integration with user-service

+++ src/main/java/com/bookface/postsservice/service/CommentsService.java
@@ -6,6 +6,7 @@ import com.bookface.postsservice.model.Comment;
 import com.bookface.postsservice.model.Post;
 import com.bookface.postsservice.repository.CommentsRepository;
 import com.bookface.postsservice.repository.PostsRepository;
+import jakarta.servlet.http.HttpServletRequest;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -29,20 +30,39 @@ import org.springframework.data.mongodb.core.query.Query;
 public class CommentsService {
     private final CommentsRepository commentsRepository;
     private final PostsRepository  postRepository;
+    private final jwtService jwtService;
 
+    public boolean compareUsername(HttpServletRequest request, String authorid){
+        String authorizationHeader = request.getHeader("Authorization");
+        // String encodedCredentials = authorizationHeader.substring(7);
+        String token=authorizationHeader.substring(7);
+        String username=jwtService.extractUsername(token);
+        return authorid.equals(username);
+
+    }
+
+    public String getuser(HttpServletRequest request){
+        String authorizationHeader = request.getHeader("Authorization");
+        // String encodedCredentials = authorizationHeader.substring(7);
+        String token=authorizationHeader.substring(7);
+        String username=jwtService.extractUsername(token);
+        return username;
+
+    }
     @CacheEvict(value = "commentCache", key="#postId")
-    public void createComment(CommentRequest commentRequest,String postId) throws Exception{
+    public void createComment(CommentRequest commentRequest,String postId,HttpServletRequest request) throws Exception{
+        String username=getuser(request);
         Optional<Post> verify_post = postRepository.findById(postId);
         if (verify_post==null){
             throw new Exception("comment can't be added wrong post id");
-    }
+        }
         if(commentRequest.getContent().length()==0){
             throw new Exception("you can't post an empty comment");
         }
 
         Comment comment=Comment.builder()
                 .postid(postId)
-                .author(commentRequest.getAuthor())
+                .author(username)
                 .content(commentRequest.getContent())
                 .numb_likes(0)
                 .createdAt(java.time.LocalDateTime.now())
@@ -56,49 +76,28 @@ public class CommentsService {
     public List<CommentResponse> getALLComments(String postId) {
         List<Comment> comments=getAllCommentsByPostId(postId);
 
-      return comments.stream().map(this::mapToCommentResponse).toList();
-       // return commentsRepository.findByPostId(postId);
+        return comments.stream().map(this::mapToCommentResponse).toList();
+        // return commentsRepository.findByPostId(postId);
 
     }
     @Autowired
     private MongoTemplate mongoTemplate;
     public List<Comment> getAllCommentsByPostId(String postId){
-
-//        Query query=new Query(Criteria.where("post_id").is(postId));
-//       List<Comment> comments =  mongoTemplate.find(query, Comment.class);
-       // List<Comment> comments = commentsRepository.findBy()
-//
-//        List<Comment> comments = mongoTemplate.find(
-//                Query.query(Criteria.where("post_id").is(postId)),
-//                Comment.class
-//        );
-
         return commentsRepository.findBypostid(postId);
-       // System.out.println(comments);
-
-       // return comments;
     }
-    /*
-    public List<Comment> getAllCommentsOfPost(String postId){
-        List<Comment> comments=commentsRepository.findAll();
-        List<Comment> returned_list=new ArrayList<>();
-        for(int i=0;i<comments.size();i++){
-            if(comments.get(i).getPost_id().equals(postId)){
-                  returned_list.add(comments.get(i));
-            }
-
-        }
-       return returned_list;
-    }*/
     public CommentResponse getCommentById(String id) {
         Optional<Comment> comments=commentsRepository.findById(id);
         return comments.map(this::mapToCommentResponse).get();
     }
     @CacheEvict(value = "commentCache", key = "#id")
-    public void updateComment(String id, String newContent) {
+    public void updateComment(String id, String newContent,HttpServletRequest request) throws Exception {
         Comment comment = commentsRepository.findById(id).orElse(null);
 
         if (comment != null) {
+            boolean check =compareUsername(request,comment.getAuthor());
+            if(!check){
+                throw new Exception("Wrong user");
+            }
             if (newContent != null && newContent.length() != 0) {
                 comment.setContent(newContent);
             }
@@ -108,24 +107,22 @@ public class CommentsService {
         }
     }
     @CacheEvict(value = "commentCache", key = "#postId")
-        public void deleteComment(String id, String postId) {
-            Comment comment = commentsRepository.findById(id).orElse(null);
-            if(comment != null) {
-                commentsRepository.deleteById(id);
-                decrementComments(postId);
+    public void deleteComment(String id, String postId,HttpServletRequest request) throws Exception {
+        // String username=getuser(request);
+        Comment comment = commentsRepository.findById(id).orElse(null);
+        if(comment != null) {
+            boolean check =compareUsername(request,comment.getAuthor());
+            if(!check){
+                throw new Exception("Wrong user");
             }
+            commentsRepository.deleteById(id);
+            decrementComments(postId);
         }
+    }
     @CacheEvict(value = "commentCache", key = "#postId")
-        public void deleteAllCommentsByPostId(String postId){
-           commentsRepository.deleteBypostid(postId);
-            /*
-           for(int i=0;i<commentsToDelete.size();i++){
-               commentsRepository.deleteById(commentsToDelete.get(i).getId());
-           }*/
-
-          //  return
-
-        }
+    public void deleteAllCommentsByPostId(String postId){
+        commentsRepository.deleteBypostid(postId);
+    }
     public void incrementLikes(String id) {
         Comment comment = commentsRepository.findById(id).orElse(null);
         if (comment!= null) {
@@ -135,7 +132,7 @@ public class CommentsService {
     }
 
     public void decrementLikes(String id) {
-         Comment comment= commentsRepository.findById(id).orElse(null);
+        Comment comment= commentsRepository.findById(id).orElse(null);
         if (comment != null) {
             int LikesCount = comment.getNumb_likes();
             if (LikesCount > 0) {
@@ -174,5 +171,5 @@ public class CommentsService {
             }
         }
     }
-    }
+}
